+++ ServerJars/ClientApi.cs
@@ -1,5 +1,7 @@
 ﻿using ServerJarsAPI.Converter;
+using ServerJarsAPI.Events;
 using ServerJarsAPI.Models;
+using System.Net.Http.Handlers;
 using System.Net.Http.Json;
 using System.Text.Json;
 
@@ -7,20 +9,29 @@ namespace ServerJarsAPI;
 
 public abstract class ClientApi : IDisposable
 {
-    protected readonly HttpClient _httpClient;
-    private readonly bool _disposeClient = true;
-    readonly JsonSerializerOptions _jsonOptions = new();
+    private readonly HttpClient _httpClient;
+    private readonly ProgressMessageHandler _progressMessageHandler;
 
-    protected ClientApi(string baseUri, HttpClient httpClient, bool disposeClient = false)
+    private readonly JsonSerializerOptions _jsonOptions = new();
+
+    protected ClientApi(string baseUri)
     {
-        _httpClient = httpClient;
-        _httpClient.BaseAddress = new Uri(baseUri);
-        _disposeClient = disposeClient;
+        _progressMessageHandler = new ProgressMessageHandler(new HttpClientHandler()
+        {
+            AllowAutoRedirect = true
+        });
+
+        _httpClient = new HttpClient(_progressMessageHandler)
+        {
+            BaseAddress = new Uri(baseUri)
+        };
 
         _jsonOptions.Converters.Add(new UnixEpochDateTimeConverter());
     }
 
-    protected async Task<T> GetAsync<T>(string uri, CancellationToken cancellationToken = default)
+    protected async Task<T> GetAsync<T>(
+        string uri,
+        CancellationToken cancellationToken = default)
     {
         using var requestMessage = new HttpRequestMessage(HttpMethod.Get, uri);
         requestMessage.Headers.Add("Accept", "application/json");
@@ -43,12 +54,14 @@ public abstract class ClientApi : IDisposable
 
     }
 
-    protected async Task<Stream> StreamAsync(string uri, CancellationToken cancellationToken = default)
+    protected async Task<Stream> StreamAsync(
+        string uri,
+        CancellationToken cancellationToken = default)
     {
         using var requestMessage = new HttpRequestMessage(HttpMethod.Get, uri);
         requestMessage.Headers.Add("Accept", "application/json");
 
-        var httpResponse = await _httpClient.SendAsync(requestMessage, cancellationToken);
+        var httpResponse = await _httpClient.SendAsync(requestMessage, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
 
         if (httpResponse.IsSuccessStatusCode)
         {
@@ -59,7 +72,42 @@ public abstract class ClientApi : IDisposable
         throw new HttpRequestException($"{httpResponse.ReasonPhrase}: {errorMsg}", null, httpResponse.StatusCode);
     }
 
-    private static async Task<string> GetErrorRepsonse(HttpResponseMessage response, CancellationToken cancellationToken = default)
+    protected async Task DownloadAsync(
+        Stream stream,
+        string uri,
+        IProgress<ProgressEventArgs>? progress,
+        CancellationToken cancellationToken = default)
+    {
+        void processEventHandler(object? _, HttpProgressEventArgs args)
+        {
+            progress?.Report(new ProgressEventArgs(args.ProgressPercentage, args.BytesTransferred, args.TotalBytes));
+        }
+
+        _progressMessageHandler.HttpReceiveProgress += processEventHandler;
+        progress?.Report(new ProgressEventArgs(0, 0));
+
+        using var requestMessage = new HttpRequestMessage(HttpMethod.Get, uri);
+        requestMessage.Headers.Add("Accept", "application/json");
+
+        using var httpResponse = await _httpClient.SendAsync(requestMessage, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
+
+        if (!httpResponse.IsSuccessStatusCode)
+        {
+            string errorMsg = await GetErrorRepsonse(httpResponse, cancellationToken);
+            throw new HttpRequestException($"{httpResponse.ReasonPhrase}: {errorMsg}", null, httpResponse.StatusCode);
+        }
+
+        progress?.Report(new ProgressEventArgs(0, 0, httpResponse.Content.Headers.ContentLength));
+
+        using var httpStream = await httpResponse.Content.ReadAsStreamAsync(cancellationToken);
+        await httpStream.CopyToAsync(stream, cancellationToken);
+
+        _progressMessageHandler.HttpReceiveProgress -= processEventHandler;
+    }
+
+    private static async Task<string> GetErrorRepsonse(
+        HttpResponseMessage response,
+        CancellationToken cancellationToken = default)
     {
         string errorMsg = string.Empty;
         try
@@ -82,10 +130,8 @@ public abstract class ClientApi : IDisposable
     {
         if (disposing)
         {
-            if (_disposeClient)
-            {
-                _httpClient?.Dispose();
-            }
+            _progressMessageHandler?.Dispose();
+            _httpClient?.Dispose();
         }
     }
 }